Extract browsePages utility function, add constant
