+++ apps/explorer/src/utils/api.ts
@@ -32,3 +32,24 @@ type QueriesCollection<T> = {
 
 export const createQueriesCollection = <T extends Record<string, Queries>>(collection: T): QueriesCollection<T> =>
   collection
+
+export const browsePages = async (
+  callback: (arg: any, options: { limit: number; page: number }) => Promise<any>,
+  pageLimit: number
+): Promise<any[]> => {
+  let pageTotalResults
+  let page = 1
+
+  const results = []
+
+  while (pageTotalResults === undefined || pageTotalResults === pageLimit) {
+    const pageResults = await callback(callback.arguments[0], { limit: pageLimit, page })
+
+    results.push(...pageResults)
+
+    pageTotalResults = pageResults.length
+    page += 1
+  }
+
+  return results
+}
