+++ packages/shared/lib/index.ts
@@ -25,5 +25,5 @@ export * from './numbers'
 export * from './constants'
 export * from './transactions'
 export * from './password-crypto'
-export * from './tokens'
 export * from './walletConnect'
+export * from './api'
